UserService/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
psycopg2-binary==2.9.5
pydantic==2.6.4
pydantic_core==2.16.3
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]